pyproject.toml
[tool.poetry]
name = "jdbottester"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
jishaku = {git = "https://github.com/Gorialis/jishaku.git"}
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
